fix: fix type error on ordering (#169)

+++ src/Label/LabelOptionsResolver.php
@@ -143,7 +143,8 @@ class LabelOptionsResolver
         if ($deliveryOptions->getShipmentOptions()) {
             $insuredAmount = $deliveryOptions->getShipmentOptions()->getInsurance();
         }
-        $insuredAmount = min($insuredAmount ?? $grandTotal, $maxAmount ?? Constant::INSURANCE_CONFIGURATION_NONE);
+
+        $insuredAmount = (int) min($insuredAmount ?? $grandTotal, $maxAmount ?? Constant::INSURANCE_CONFIGURATION_NONE);
 
         return $this->getHighestAllowedValue($insuredAmount, $consignment->getInsurancePossibilities());
     }
